refactor vscode part; seem to work

+++ crates/domains/ml/datasets/src/labeled.rs
@@ -3,5 +3,5 @@ use vm::AnyValueDyn;
 
 pub struct LabeledData<'eval> {
     pub input: Arc<dyn AnyValueDyn<'eval>>,
-    pub label: i32,
+    pub label: u8,
 }
